requirements.txt
Flask
tensorflow
numpy
h5py
Pillow
